requirements.txt
arxiv
sentence-transformers
faiss-cpu
pandas
numpy
tqdm
supabase
python-dotenv
